rename: d3d12: pool and descriptor set
