+++ Project/D3D12/v16/Buma3D.vcxproj
@@ -100,7 +100,9 @@
     <ClInclude Include="..\..\..\Source\API\D3D12\DebugMessageD3D12.h" />
     <ClInclude Include="..\..\..\Source\API\D3D12\DebugMessageQueueD3D12.h" />
     <ClInclude Include="..\..\..\Source\API\D3D12\DepthStencilViewD3D12.h" />
+    <ClInclude Include="..\..\..\Source\API\D3D12\DescriptorPool0D3D12.h" />
     <ClInclude Include="..\..\..\Source\API\D3D12\DescriptorPoolD3D12.h" />
+    <ClInclude Include="..\..\..\Source\API\D3D12\DescriptorSet0D3D12.h" />
     <ClInclude Include="..\..\..\Source\API\D3D12\DescriptorSetD3D12.h" />
     <ClInclude Include="..\..\..\Source\API\D3D12\DescriptorSetLayoutD3D12.h" />
     <ClInclude Include="..\..\..\Source\API\D3D12\DeviceAdapterD3D12.h" />
@@ -151,7 +153,9 @@
     <ClCompile Include="..\..\..\Source\API\D3D12\DebugMessageD3D12.cpp" />
     <ClCompile Include="..\..\..\Source\API\D3D12\DebugMessageQueueD3D12.cpp" />
     <ClCompile Include="..\..\..\Source\API\D3D12\DepthStencilViewD3D12.cpp" />
+    <ClCompile Include="..\..\..\Source\API\D3D12\DescriptorPool0D3D12.cpp" />
     <ClCompile Include="..\..\..\Source\API\D3D12\DescriptorPoolD3D12.cpp" />
+    <ClCompile Include="..\..\..\Source\API\D3D12\DescriptorSet0D3D12.cpp" />
     <ClCompile Include="..\..\..\Source\API\D3D12\DescriptorSetD3D12.cpp" />
     <ClCompile Include="..\..\..\Source\API\D3D12\DescriptorSetLayoutD3D12.cpp" />
     <ClCompile Include="..\..\..\Source\API\D3D12\DeviceAdapterD3D12.cpp" />
