+++ Project/D3D12/v16/Buma3D_D3D12_DLL.vcxproj.filters
@@ -51,6 +51,9 @@
     <Filter Include="Source\D3D12\Query">
       <UniqueIdentifier>{3da7fa12-f3a8-428c-a0d6-e68326842a7f}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Source\D3D12\Descriptor\Deprecated">
+      <UniqueIdentifier>{32b2b433-7311-4cf3-a647-1dadb189bf84}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\..\Source\API\D3D12\DeviceFactoryD3D12.cpp">
@@ -175,6 +178,18 @@
     </ClCompile>
     <ClCompile Include="..\..\..\Source\API\D3D12\Buma3DDLLMain.cpp">
       <Filter>Source</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\Source\API\D3D12\DescriptorPool0D3D12.cpp">
+      <Filter>Source\D3D12\Descriptor\Deprecated</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\Source\API\D3D12\DescriptorSet0D3D12.cpp">
+      <Filter>Source\D3D12\Descriptor\Deprecated</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\Source\API\D3D12\DescriptorSetLayoutD3D12.cpp">
+      <Filter>Source\D3D12\Descriptor</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\..\Source\API\D3D12\PipelineLayoutD3D12.cpp">
+      <Filter>Source\D3D12\Descriptor</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -321,6 +336,18 @@
     </ClInclude>
     <ClInclude Include="..\..\..\Source\API\D3D12\StreamOutputBufferViewD3D12.h">
       <Filter>Source\D3D12\View</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\Source\API\D3D12\DescriptorPool0D3D12.h">
+      <Filter>Source\D3D12\Descriptor\Deprecated</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\Source\API\D3D12\DescriptorSet0D3D12.h">
+      <Filter>Source\D3D12\Descriptor\Deprecated</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\Source\API\D3D12\PipelineLayoutD3D12.h">
+      <Filter>Source\D3D12\Descriptor</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\Source\API\D3D12\DescriptorSetLayoutD3D12.h">
+      <Filter>Source\D3D12\Descriptor</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
